Refactoring and remove unused
