+++ src/netspresso_trainer/models/heads/detection/experimental/detection/__init__.py
@@ -1,5 +1,5 @@
-from .anchor_utils import *
 from .roi_heads import *
 from .rpn import *
 from .generalized_rcnn import *
 from .roi_align import *
+from ._utils import *
