--- src/netspresso_trainer/models/heads/detection/experimental/detection/anchor_utils.py
@@ -1,133 +1,0 @@
-from typing import List
-
-import torch
-from torch import Tensor, nn
-
-
-class AnchorGenerator(nn.Module):
-    """
-    Module that generates anchors for a set of feature maps and
-    image sizes.
-
-    The module support computing anchors at multiple sizes and aspect ratios
-    per feature map. This module assumes aspect ratio = height / width for
-    each anchor.
-
-    sizes and aspect_ratios should have the same number of elements, and it should
-    correspond to the number of feature maps.
-
-    sizes[i] and aspect_ratios[i] can have an arbitrary number of elements,
-    and AnchorGenerator will output a set of sizes[i] * aspect_ratios[i] anchors
-    per spatial location for feature map i.
-
-    Args:
-        sizes (Tuple[Tuple[int]]):
-        aspect_ratios (Tuple[Tuple[float]]):
-    """
-
-    __annotations__ = {
-        "cell_anchors": List[torch.Tensor],
-    }
-
-    def __init__(
-        self,
-        sizes=((128, 256, 512),),
-        aspect_ratios=((0.5, 1.0, 2.0),),
-        image_size=(512, 512),
-    ):
-        super().__init__()
-
-        if not isinstance(sizes[0], (list, tuple)):
-            # TODO change this
-            sizes = tuple((s,) for s in sizes)
-        if not isinstance(aspect_ratios[0], (list, tuple)):
-            aspect_ratios = (aspect_ratios,) * len(sizes)
-
-        self.sizes = sizes
-        self.aspect_ratios = aspect_ratios
-        self.cell_anchors = [
-            self.generate_anchors(size, aspect_ratio) for size, aspect_ratio in zip(sizes, aspect_ratios)
-        ]
-        
-        self.image_size = image_size
-
-    # TODO: https://github.com/pytorch/pytorch/issues/26792
-    # For every (aspect_ratios, scales) combination, output a zero-centered anchor with those values.
-    # (scales, aspect_ratios) are usually an element of zip(self.scales, self.aspect_ratios)
-    # This method assumes aspect ratio = height / width for an anchor.
-    def generate_anchors(
-        self,
-        scales: List[int],
-        aspect_ratios: List[float],
-        dtype: torch.dtype = torch.float32,
-        device: torch.device = torch.device("cpu"),
-    ) -> Tensor:
-        scales = torch.as_tensor(scales, dtype=dtype, device=device)
-        aspect_ratios = torch.as_tensor(aspect_ratios, dtype=dtype, device=device)
-        h_ratios = torch.sqrt(aspect_ratios)
-        w_ratios = 1 / h_ratios
-
-        ws = (w_ratios[:, None] * scales[None, :]).view(-1)
-        hs = (h_ratios[:, None] * scales[None, :]).view(-1)
-
-        base_anchors = torch.stack([-ws, -hs, ws, hs], dim=1) / 2
-        return base_anchors.round()
-
-    def set_cell_anchors(self, dtype: torch.dtype, device: torch.device):
-        return [cell_anchor.to(dtype).to(device) for cell_anchor in self.cell_anchors]
-
-    def num_anchors_per_location(self) -> List[int]:
-        return [len(s) * len(a) for s, a in zip(self.sizes, self.aspect_ratios)]
-
-    # For every combination of (a, (g, s), i) in (self.cell_anchors, zip(grid_sizes, strides), 0:2),
-    # output g[i] anchors that are s[i] distance apart in direction i, with the same dimensions as a.
-    def grid_anchors(self, cell_anchors, grid_templates, grid_sizes: List[List[int]], strides: List[List[Tensor]]) -> List[Tensor]:
-        anchors = []
-        cell_anchors = cell_anchors
-        torch._assert(cell_anchors is not None, "cell_anchors should not be None")
-        torch._assert(
-            len(grid_sizes) == len(strides) == len(cell_anchors),
-            "Anchors should be Tuple[Tuple[int]] because each feature "
-            "map could potentially have different sizes and aspect ratios. "
-            "There needs to be a match between the number of "
-            "feature maps passed and the number of sizes / aspect ratios specified.",
-        )
-
-        for grid_template, size, stride, base_anchors in zip(grid_templates, grid_sizes, strides, cell_anchors):
-            grid_height_template, grid_width_template = grid_template
-            grid_height, grid_width = size
-            stride_height, stride_width = stride
-            device = base_anchors.device
-
-            # For output anchor, compute [x_center, y_center, x_center, y_center]
-            #shifts_x = torch.arange(0, grid_width, dtype=torch.int32).to(device) * stride_width
-            #shifts_y = torch.arange(0, grid_height, dtype=torch.int32).to(device) * stride_height
-            shifts_x = (torch.ones_like(grid_width_template, dtype=torch.int32).cumsum(0) - 1).to(device) * stride_width
-            shifts_y = (torch.ones_like(grid_height_template, dtype=torch.int32).cumsum(0) - 1).to(device) * stride_height
-            shift_y, shift_x = torch.meshgrid(shifts_y, shifts_x, indexing="ij")
-            shift_x = shift_x.reshape(-1)
-            shift_y = shift_y.reshape(-1)
-            shifts = torch.stack((shift_x, shift_y, shift_x, shift_y), dim=1)
-
-            # For every (base anchor, output anchor) pair,
-            # offset each zero-centered base anchor by the center of the output anchor.
-            anchors.append((shifts.view(-1, 1, 4) + base_anchors.view(1, -1, 4)).reshape(-1, 4))
-
-        return anchors
-
-    def forward(self, feature_maps: List[Tensor]) -> List[Tensor]:
-        grid_sizes = [feature_map.shape[-2:] for feature_map in feature_maps]
-        # each feature_map has (b, c, h, w) shape
-        grid_templates = [(feature_map[0, 0, :, 0], feature_map[0, 0, 0, :]) for feature_map in feature_maps]
-        dtype, device = feature_maps[0].dtype, feature_maps[0].device
-        strides = [
-            [
-                torch.empty(1, dtype=torch.int64).fill_(self.image_size[0] // g[0]).to(device),
-                torch.empty(1, dtype=torch.int64).fill_(self.image_size[1] // g[1]).to(device),
-            ]
-            for g in grid_sizes
-        ]
-        cell_anchors = self.set_cell_anchors(dtype, device)
-        anchors_over_all_feature_maps = self.grid_anchors(cell_anchors, grid_templates, grid_sizes, strides)
-
-        return anchors_over_all_feature_maps
